Move final plugin (indexing-engine) from cumulus to base project
This lets us remove the cumulus dir altogether!

Moved decision making on whether to load the module into plugins/index.
